requirements.txt
astroid==2.4.2
blueprint==3.4.2
click==7.1.1
colorama==0.4.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.0
pandas==1.1.4
psycopg2==2.8.5
pylint==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.6
pylint-sqlalchemy==0.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
scipy==1.1.0
six==1.14.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
